add stopped_at_obstacle/stopping_at_obstacle action
