+++ local_waypoint_server/config/behavior_tree.xml
@@ -1,7 +1,11 @@
 <root main_tree_to_execute = "LocalWaypointsServerTree">
     <BehaviorTree ID="LocalWaypointsServerTree">
-    <Sequence name="root_sequence">
-        <ObstacleAvoidAction name="obstacle_avoid"/>
-    </Sequence>
+    <Fallback name="root_sequence">
+        <FollowPathAction name="follwo_path"/>
+        <Sequence name="obstacle_stop_sequence">
+            <StoppingAtObstacleAction name="stopping_at_obstacle"/>
+            <StoppedAtObstacleAction name="stopped_at_obstacle"/>
+        </Sequence>
+    </Fallback>
     </BehaviorTree>
 </root>
